调整PageHelper版本

+++ assets/src/main/java/com/dili/assets/service/impl/AssetsServiceImpl.java
@@ -113,7 +113,7 @@ public class AssetsServiceImpl extends BaseServiceImpl<Assets, Long> implements 
             countInput.setParentId(null);
             countInput.setPage(null);
             countInput.setRows(null);
-            long count = this.listByExample(countInput).size();
+            int count = this.listByExample(countInput).size();
             boolean expand = false;
             if (input.getId() != null) {
                 expand = true;
@@ -123,7 +123,7 @@ public class AssetsServiceImpl extends BaseServiceImpl<Assets, Long> implements 
             BasePage<Assets> boothBasePage = this.listPageByExample(input);
             var booths = boothBasePage.getDatas();
             if (input.getParentId() == null) {
-                count = boothBasePage.getTotalItem();
+                count = boothBasePage.getTotalItem().intValue();
             }
             var results = ValueProviderUtils.buildDataByProvider(input, booths);
             var result = new ArrayList();
